Lazy internalize of JS objects
This is an hard balance to find, but converting all props, state etc. to Go objects
is a wormhole of inifinite loops and potential performance issues.

This commit revises all of this; props, state etc. are read into a Go map, but very shallow. All the values are
JS types.

This means more conversion in the client code, but this is easier to reason about than subtle blowups.